requirements.txt
langchain==0.2.15
langchain_community==0.2.14
langchain_core==0.2.36
langchain_openai==0.1.8
langchain_text_splitters==0.2.0
openai==1.31.0
python-dotenv==1.0.1
jq==1.7.0
faiss-cpu==1.8.0
xmltodict==0.13.0
langsmith==0.1.107
Flask==3.0.3
Flask-Cors==4.0.1
bs4==0.0.2
playwright==1.46.0
beautifulsoup4==4.12.3
html2text==2024.2.26
langchain-chroma==0.1.3
requests-html==0.10.0
